Changes in models
Tested the 3d printability of the models and came to the conclusion that right now, at this size of pins it's not realistic that everyone will be able to create the same results. Which means that a perfect pressure fit is very debatable for current fdm printers. Therefore the pins can only be used as a foolproofing method of inserting components into the casing.